requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.25.0
requests>=2.31.0
pydantic>=2.0.0
tenacity>=8.2.0
tiktoken>=0.5.0
rich>=13.0.0
click>=8.1.0